Lv5floorboard, Lv5icewall, obj_lv5swice, obj_lv5ychndlr   done (#2175)
* lv5floorboard done

* obj_lv5icewall done

* remove asm

* obj_lv5swice done

* obj_lv5ychndlr done

+++ include/d/a/d_a_alink.h
@@ -3078,7 +3078,7 @@ public:
     virtual bool checkSpinnerPathMove();
     virtual bool checkSpinnerTriggerAttack();
     virtual void onSpinnerPathForceRemove();
-    virtual s16 getIronBallBgHit() const;
+    virtual int getIronBallBgHit() const;
     virtual cXyz* getIronBallCenterPos();
     virtual bool checkCanoeFishingGetLeft() const;
     virtual bool checkCanoeFishingGetRight() const;
